[REXA-21] Add ControllerAdviceHandler to handle errors (#6)
# Context
In this PR, I am adding a ControllerAdviceHandler and some lines of code to handle generic errors from Spring. We can later enrich it with other errors to return a ApiErrorBean to the user when using our APIs.

# Extra
I added as well a new plugin to generate our dependencies in maven (as well as in the front). I tried to clean as much as possible all of them.