[Rexa 5] Add user roles and management for ADMIN (#25)
# Context

In this PR, I am adding two roles : 
- ADMIN
- USER

Depending on the roles, you can see a new page to manage your user. You can disable or enable any user that are not administrator.

# Fixes

I also edited all the endpoints for the front. I added the annotation `@Secured` with ADMIN or USER or both depending on the endpoint.

* [REXA-5] Review + fix some issues